Force all threads to sleep until the throttle is released
Previously it looks like Quip was consistently returning a 503 for
throttled requests, but now it also returns a 429. The 429 does not
return throttle information - causing those requests to fail before
the limits reset. With this change when we encounter a 503 with a
x-ratelimit-reset time we force all threads to sleep until this time.